% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztcoins.R
\name{deitynumber}
\alias{deitynumber}
\title{Looks for a deity in database}
\usage{
deitynumber(deityname)
}
\arguments{
\item{deityname}{takes a string in quotes}
}
\value{
the total number of coins featuring a particular deity
}
\description{
Searches a coin database for the total number of coins featuring deity
}
\examples{
deitynumber("Venus")

}
