% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztcoins.R
\name{denominations}
\alias{denominations}
\title{Plot the denominations of coins}
\usage{
denominations()
}
\arguments{
\item{none}{}
}
\value{
A bar plot of the results
}
\description{
Computes the number of each type of denomination in a database as a bar plot
}
\examples{
denominations()

}
