% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztcoins.R
\name{denomnumber}
\alias{denomnumber}
\title{Looks for a denomination type in database}
\usage{
denomnumber(denom)
}
\arguments{
\item{denom}{takes a string in quotes}
}
\value{
the total number of coins of chosen denomination
}
\description{
Searches a coin database for the total number of coins of a particular denomination
}
\examples{
denomnumber("Denarius")

}
