% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztcoins.R
\name{emperors}
\alias{emperors}
\title{Plot the emperors appearing on coins}
\usage{
emperors()
}
\arguments{
\item{none}{}
}
\value{
A bar plot of the results
}
\description{
Computes the number of times an emperor appears on coins in the database as a bar plot
}
\examples{
emperors()

}
