% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztcoins.R
\name{empnumber}
\alias{empnumber}
\title{Looks for emeperor in database}
\usage{
empnumber(empname)
}
\arguments{
\item{empname}{takes a string in quotes}
}
\value{
the total number of coins featuring a particular emperor
}
\description{
Searches a coin database for the total number of coins featuring emperor
}
\examples{
empnumber("Maximian")

}
