% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztcoins.R
\name{materials}
\alias{materials}
\title{Plot the coins minted in each material}
\usage{
materials()
}
\arguments{
\item{none}{}
}
\value{
A bar plot of the results
}
\description{
Computes the number of coins minted in each material and plots it as a bar plot
}
\examples{
materials()

}
