% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztcoins.R
\name{matnumber}
\alias{matnumber}
\title{Looks for a material type in database}
\usage{
matnumber(mat)
}
\arguments{
\item{mat}{takes a string in quotes}
}
\value{
the total number of coins featuring made from a particular material
}
\description{
Searches a coin database for the total number of coins made from a material
}
\examples{
matnumber("Gold")

}
