% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztcoins.R
\name{mintnumber}
\alias{mintnumber}
\title{Looks for a mint location in database}
\usage{
mintnumber(mint)
}
\arguments{
\item{mint}{takes a string in quotes}
}
\value{
the total number of coins minted at a chosen city
}
\description{
Searches a coin database for the total number of coins minted at a city
}
\examples{
mintnumber("Trier")

}
