% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztcoins.R
\name{mints}
\alias{mints}
\title{Plot the number of coins minted in each mint}
\usage{
mints()
}
\arguments{
\item{none}{}
}
\value{
A bar plot of the results
}
\description{
Computes the number of coins minted at each mint and plots it as a bar plot
}
\examples{
mints()

}
