% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztcoins.R
\name{regionnumber}
\alias{regionnumber}
\title{Looks for a region in database}
\usage{
regionnumber(region)
}
\arguments{
\item{region}{takes a string in quotes}
}
\value{
the total number of coins from a chosen region
}
\description{
Searches a coin database for the total number of coins minted in a region
}
\examples{
regionnumber("Gallia")

}
