% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztcoins.R
\name{yearnumber}
\alias{yearnumber}
\title{Looks for a year in database}
\usage{
yearnumber(year)
}
\arguments{
\item{year}{takes a string in quotes}
}
\value{
the total number of coins minted in a chosen year
}
\description{
Searches a coin database for the total number of coins minted in a year
}
\examples{
yearnumber("145")

}
